[FIX] project: improve kanban view of task
Purpose of this commit improve kanban_state field layout in
kanban view of task.when user_id doesn't set on task it show
blank div for m2oAvatar which move kanban_state field to left
of that div.

So, In this commit show m2oAvatar only when use_id
is set on that task.

closes odoo/odoo#75258

Taskid: 2557892
X-original-commit: 7e8daad39b44a1ffd7e2c62b2125afba0fe3b962
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>