[FIX] sale: pass invoice origin as string instead of list
Currently, the default invoice origin for the sale action_view_invoice
is being passed as a list by using the mapped method. This causes that,
when an invoice is created from this view, the origin is set to ['SOXXXXX'].

With the fix, we properly pass the origin as a string so the origin is set
as SOXXXXX, which should be the expected. Notice also that the context is
only passed when dealing with a single SO.

closes odoo/odoo#74579

X-original-commit: 32acaa88c1fdb2b735a3defe67c32a1a5c1e4b96
Signed-off-by: Victor Feyens (vfe) <vfe@odoo.com>