[FIX] account: payment method line deletion/constrains
Set the payment method line to cascade when deleting a journal.
This way, we do not keep unneeded lines, and it is not an issue
since we cannot delete journals linked to a payment/move anyway.

Also add a check on the journal id in the constrains checking their
name to avoid taking into account lines without journals.

closes odoo/odoo#74767

X-original-commit: 918513e8d80f2be733f02834867d28c6b5f13b38
Signed-off-by: William André (wan) <wan@odoo.com>