[FIX] payment_adyen: always reply with `'[accepted]'` to notifications
Before this commit, when the webhook was processing a notification,
if there was a validation error, the webhook was sending the whole
traceback to Adyen. Since this was still a response code 200, this
didn't cause any problem, but we could see the response received in
the Adyen backend.

So we will now catch any validation error and send `'[accepted]'` in the
response.

closes odoo/odoo#74733

Signed-off-by: Antoine Vandevenne (anv) <AntoineVDV@users.noreply.github.com>