[FIX] google_calendar: prevent weird behavior when partner email and user email differs.
This commit allows using different mail on the user and the partner without experiencing weird behaviors.

Before this commit, sometimes when user would set a different email on his user_id and his main partner, the sync event would not be visible as long as the event was not accepted in Google.

closes odoo/odoo#74903

X-original-commit: d60fead9bb331f4657365d916905c757ac26bad9
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>